// MCP_Frontend-main/src/app/shared/models/user-data.ts
export class UserData {
    employeeId: number;
    userName: string;
    fullName: string;
    userRoleId: string;
    emailAddress: string;
    phoneNo: string;
    receiveBillerStmnt: string;
    lockUser: string;
    pwdEnhancedFlag: string;
    userPwd: string;
    status: string;
    createdBy: string;
    createDt: Date;
    modifyBy: string;
    modifyDt: Date;
    edit: Boolean;
}